dcr: Avoid freeze inside db transaction.
In dcr there's a request for the current wallet inside a db
transaction. If a user closes a wallet during
*asset.IndexTransactions() It will get the lock at
*dcrLoader.UnloadWallet() but then get stuck at l.db.Close() because it
cannot close the db while a transation is taking place. So the request
in the transation gets stuck trying to get the lock. Prevent that by
holding a db lock while indexing transactions that stops the wallet
from being unloaded during that time.